requirements.txt
fastapi
uvicorn
pydantic
adafruit-io
pydantic

python-dotenv

websockets

pymongo
certifi
asyncpg

pytest
